reached point where need to make solid decisions on constraints impl
at first just added constraints to unknowns, made it easy to get the
basic code down.

You can `declare` that some unknown satisfies a constraint and thsi
propegates with the unknown.

Unifying two unknowns takes the union of the constraints.

A constraint is just some other type.

'known' types can't have constraints, a constraint it purely a
requirement on an unknown...which kind of matches the c# style of
contraining a generic type param.

Now I need to know if a type satisfies a constraint and I'm not sure
how to expose that. Do we want a specific kind of type that constaints
must be? or maybe we expose some 'satisfies-p' function the user can
implement.

I do need some way to specify if something isnt a valid thing to use
as a constraint. What would that be exactly? I mean we could have a
default behaviour where satifying means the types are the same (or
unify I guess).. if we use unify we need to protect against side
effects.

Clearly the goal is to allow different kinds of type systems but not
to make a soupy mess. When we implement traits we want to do so using
constraints. The trait itself will be a ttype, but there will be no
compiled respresentation for it, and (given not cockups on the user's
part) there would be no function that returns a trait as a type.. it
might be nice to let users restrict things though so you can get
better errors. hmm.

Whats the best way of doing that? could just store a bool saying its a
constrait-only type, we would then error if any function tries to take
one as  an arg or tries to return one.

That could work :)

what about the converse case, what to ensure it is never used as a
constraint? could use the same flag I guess, just make it a
keyword. so the types 'purpose' could be: :constraint-only,
:concrete-only or :concreate-and-constraint... that might be enough
actually.

let's try it