refactor to mutate graph and have direct refs to bindings
The direct reference to bindings means we dont have to make an
environment to store bindings every time we walk the tree.

This also pushes it a bit more towards the sea of nodes approach as it
essentailly is a simple dependency graph.

The mutation helps make maintaining the references easy. It was a mild
concern at first but the passes are so simple that its easy to see the
logic.