treat #'function forms as constants when optimizing
This was meant to be temporary whilst we sort out inline of top level
functions, however this kinda makes sense. A specific pass could
inline these later.

It's nice to see that the code does optimize as far as it can with the
current passes

    #(SSAD-LET1 ((G1012 (SSAD-FUNCALL (CONSTANT #'+)
                                      (CONSTANT 20)
                                      (CONSTANT 20))))
      (VAR G1012))

With top level func inlining and compiler macros all this is going
away :)