swap user-constant-folds for optimize-macros
These are passed the value (if constant) or an opaque value
handle. You are then allowed to return a form which we be compiled and
may contain those opaque values. The user can use var-eq to compare to
see if they are refs to the same binding

The opaque values are just instances of ssad-var. Infer is tweaked to
splice them.

I'm not satisfied that this is really the way to go, but it does
expose the tranform to the user which is nice. The issues are that
because we have to accept list forms we have to compile and that's
less efficient than it could be. Also as you only have constants and
vars you can't see places where associativity and such would help
you. I am not sure if we will expose this or what interface would make
sense.