rename dummy-funcs, add wrap-sized, add define-backend, add emit
Getting everything set up to be able to emit lisp code. For this we
need a backend which will hold the value-type info and the op function
emitters.

We have it working for our little example. It's kinda odd that we
require a central repository of op-expanders but then force emit to
belong to the backend.. maybe more sensible to let each backend handle
it. That would make it harder to extend a backend of course..that
could be undesirable.