add slot-value special form and update everything to use it
We tranfrom record slot access into slot-value calls. This lets the
backends decided how they want to handle record access. This is
important as soon enough we will be adding layouts which will serious
effect how records are represented in the code.

I tried to get a simple query using records to compile but it chokes
on record constructors right now. We'll fix that soon.

We need to have another think about when layout comes into play. If we
knew earlie that the results were going into seperate columns we could
handle this in the first passes. The thing is it is really about
tables so it would probably be premature to do that at first... then
again we really need to work out what info is known when we start
compiling these things. Probably will be much clearer when we start
writing the query-set code.